Updated how an application's media is copied.
It was copied last, but then it would never get copied again unless you purged it. Now it copies it if it has changed.

Introduces a slight bug: If the media of the project overrides an application's media, the file will constantly be replaced: first the application will think it is out of date and copy it, then the project will do the same. Not sure of a solution yet.

Also fixes a bug in combining files